Day86_TypingSpeedTest/requirements.txt
request==2.31.0